Use `int` coordinates in `Sprite::draw`
Now that we delegate to `AnimatedImage::draw`, which uses `int` coordinates, there is now auto conversion to `int`. We might as well declare `Sprite` parameters to be `int` based as well.

+++ NAS2D/Resource/Sprite.cpp
@@ -93,13 +93,13 @@ void Sprite::update()
 }
 
 
-void Sprite::draw(Point<float> position) const
+void Sprite::draw(Point<int> position) const
 {
 	mAnimatedImage.draw(Utility<Renderer>::get(), position, mTintColor);
 }
 
 
-void Sprite::draw(Point<float> position, Angle rotation) const
+void Sprite::draw(Point<int> position, Angle rotation) const
 {
 	mAnimatedImage.draw(Utility<Renderer>::get(), position, mTintColor, rotation);
 }
